feat : 무기 MVP 완료
근접공격(플레이어용)
단일 공격(MG)
폭발 공격(바주카)
레이캐스트 공격(빔 라이플) 구현

+++ Assets/Gundam/Game/SO/NPC/BastStatus/BaseStatueSO.cs
@@ -6,6 +6,6 @@ public class BaseStatueSO : ScriptableObject
 {
     public float walkSpeed;
     public float runSpeed;
-    public float maxHp;
+    public int maxHp;
 }
